refactor(tests): consolidate static import tracing utilities for better reuse
