+++ apps/app/src/features/mastra/server/test-utils/index.ts
@@ -1,0 +1,5 @@
+export type { StaticImportTraceInput } from './trace-static-import-chains';
+export {
+  listMissingEntrypoints,
+  traceForbiddenPackageChains,
+} from './trace-static-import-chains';
